:recycle: navigation bar 리팩토링 #4 - hook에서 컴포넌트로 변경 - navigation menu 컴포넌트 삭제 - navigation config를 shared에서 widget/header로 이동 - navigation bar를 logo bar로 이름 변경
