+++ src/containers/(account)/index.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 
-import { useNavigationBar } from '@/features/nav-bar/hook';
+import Logo from '@/features/logo-bar';
 
 import styles from './index.module.scss';
 
@@ -9,12 +9,10 @@ interface Props {
 }
 
 const AccountLayout = async ({ children }: Props) => {
-  const { NavigationBar } = useNavigationBar();
-
   return (
     <>
       <header className='max-md:px-2 xl:pt-8 md:py-1 xl:drop-shadow-sm'>
-        <NavigationBar />
+        <Logo business={false} />
       </header>
       <main className={styles.main}>
         <div className={styles.container}>{children}</div>
