+++ src/features/logo-bar/index.ts
@@ -1,0 +1,1 @@
+export { default } from '@/features/logo-bar/ui';
