+++ src/features/logo-bar/ui/logo.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 
 interface Props {
-  business?: boolean;
+  business: boolean;
 }
 
 const Logo = ({ business }: Props) => {
