--- src/features/nav-bar/index.ts
@@ -1,1 +1,0 @@
-export { useNavigationBar } from '@/features/nav-bar/hook';
