+++ src/shared/config/index.ts
@@ -2,5 +2,3 @@ export const { BASE_URL } = process.env;
 
 export const DOMAIN_NAME = process.env.DOMAIN_NAME || '/error';
 export const BIZ_DOMAIN_NAME = process.env.BIZ_DOMAIN_NAME || '/biz';
-
-export { NAVIGATIONS } from './navigation';
