+++ src/widgets/header/config/navigation.ts
