+++ src/widgets/header/ui/index.tsx
@@ -1,23 +1,39 @@
+import Link from 'next/link';
+
 import { Drawer, DrawerClose, DrawerOpen } from '@/features/drawer';
 import Login from '@/features/login';
-import { useNavigationBar } from '@/features/nav-bar';
+import NavigationBar from '@/features/logo-bar';
 
 import DrawerLogin from './drawer-item/login';
 import NaviationBelt from './nav-belt';
+import { NAVIGATIONS } from '../config/navigation';
 
 interface Props {
   business?: boolean;
 }
 
 const Header = ({ business }: Props) => {
-  const { NavigationBar, NavigationMenu } = useNavigationBar(business);
+  const isBusiness = !!business;
+  const navigationMenuItems = business
+    ? NAVIGATIONS.business
+    : NAVIGATIONS.personal;
 
   return (
     <header>
-      <NaviationBelt business={!!business} />
-      <NavigationBar>
+      <NaviationBelt business={isBusiness} />
+      <NavigationBar business={isBusiness}>
         <span className='size-full flex items-center max-md:hidden'>
-          <NavigationMenu />
+          <nav className='h-full flex mx-auto'>
+            {navigationMenuItems.map((item) => (
+              <Link
+                className='px-4 text-sm font-medium h-full flex items-center hover:text-blue-700'
+                key={item.key}
+                href={item.path}
+              >
+                {item.name}
+              </Link>
+            ))}
+          </nav>
           <Login />
         </span>
         <DrawerOpen className='md:hidden ml-auto' />
